[Security Solution][Tech Debt] Move remaining chart related components to alert KPI folder (#223570)
## Summary

This PR moves the `chart_settings_popover` and `field_selections` to
`~security_solution/public/detections/components/alerts_kpis` since they
are used exclusively by the KPI charts. These were left out in the
original folder reorg effort in
https://github.com/elastic/kibana/pull/189234

### Checklist

- [x] The PR description includes the appropriate Release Notes section,
and the correct `release_note:*` label is applied per the
[guidelines](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)

(cherry picked from commit c04d1782b7a09bfa75cb3e3c6c4a0ef06d839013)

# Conflicts:
#	packages/kbn-babel-preset/styled_components_files.js
