+++ x-pack/solutions/security/plugins/security_solution/public/detections/components/alerts_kpis/chart_panels/chart_context_menu/index.tsx
@@ -8,8 +8,8 @@
 import { noop } from 'lodash/fp';
 import React, { useCallback } from 'react';
 
-import { ChartSettingsPopover } from '../../../../../common/components/chart_settings_popover';
-import { useChartSettingsPopoverConfiguration } from '../../../../../common/components/chart_settings_popover/configurations/default';
+import { ChartSettingsPopover } from '../chart_settings_popover';
+import { useChartSettingsPopoverConfiguration } from '../chart_settings_popover/configurations/default';
 
 interface Props {
   defaultStackByField: string;
