+++ x-pack/solutions/security/plugins/security_solution/public/detections/components/alerts_kpis/common/field_selection.test.tsx
@@ -8,13 +8,10 @@
 import { render, screen } from '@testing-library/react';
 import React from 'react';
 
-import { TestProviders } from '../../mock';
-import type { Props } from '.';
-import { FieldSelection } from '.';
-import {
-  GROUP_BY_LABEL,
-  GROUP_BY_TOP_LABEL,
-} from '../../../detections/components/alerts_kpis/common/translations';
+import { TestProviders } from '../../../../common/mock';
+import type { Props } from './field_selection';
+import { FieldSelection } from './field_selection';
+import { GROUP_BY_LABEL, GROUP_BY_TOP_LABEL } from './translations';
 
 jest.mock('react-router-dom', () => {
   const actual = jest.requireActual('react-router-dom');
